# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mindstrata API",
    "version": "2026-06-05",
    "description": "Проверяемый контракт Go API. Поле x-router-path связывает OpenAPI path с регистрацией mux.HandleFunc в apps/api/internal/httpapi/router.go."
  },
  "servers": [
    {
      "url": "https://mindstrata.ru",
      "description": "Production"
    },
    {
      "url": "https://stage.mindstrata.ru",
      "description": "Staging"
    }
  ],
...
  "paths": {
...
    "/api/admin/tariffs": {
      "x-router-path": "/api/admin/tariffs",
      "get": { "operationId": "getAdminTariffs", "tags": ["admin"], "security": [{ "cookieAuth": [] }], "responses": { "200": { "$ref": "#/components/responses/List" }, "403": { "$ref": "#/components/responses/JSONError" } } },
      "post": { "operationId": "postAdminTariffs", "tags": ["admin"], "security": [{ "cookieAuth": [] }], "requestBody": { "$ref": "#/components/requestBodies/Tariff" }, "responses": { "201": { "$ref": "#/components/responses/JSONOK" }, "400": { "$ref": "#/components/responses/JSONError" } } }
    },
...
    "/api/admin/tariff-groups": {
      "x-router-path": "/api/admin/tariff-groups",
      "get": { "operationId": "getAdminTariffGroups", "tags": ["admin"], "security": [{ "cookieAuth": [] }], "responses": { "200": { "$ref": "#/components/responses/List" }, "403": { "$ref": "#/components/responses/JSONError" } } },
      "post": { "operationId": "postAdminTariffGroups", "tags": ["admin"], "security": [{ "cookieAuth": [] }], "requestBody": { "$ref": "#/components/requestBodies/TariffGroup" }, "responses": { "201": { "$ref": "#/components/responses/JSONOK" }, "400": { "$ref": "#/components/responses/JSONError" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mindstrata_session"
      }
    },
...
    "requestBodies": {
...
      "Tariff": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TariffRequest" } } } },
      "TariffGroup": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TariffGroupRequest" } } } },
...
    },
    "responses": {
      "JSONOK": {
        "description": "JSON response with ok=true on success",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OKResponse" } } }
      },
      "JSONError": {
        "description": "JSON error response",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "List": {
        "description": "List response",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListResponse" } } }
      }
    },
    "schemas": {
      "OKResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": { "ok": { "type": "boolean", "const": true } },
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "type": "boolean", "const": false },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "errorCode": { "type": "string" },
          "requiresAuth": { "type": "boolean" }
        },
        "additionalProperties": true
      },
...
      "ListResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": {
          "ok": { "type": "boolean" },
          "items": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        },
        "additionalProperties": true
      },
...
      "TariffRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "priceRub": { "type": "integer", "minimum": 0 },
          "modeIds": { "type": "array", "items": { "type": "integer", "format": "int64" } },
          "active": { "type": "boolean" }
        },
        "additionalProperties": true
      },
      "TariffGroupRequest": {
        "type": "object",
        "required": ["name"],
        "properties": { "name": { "type": "string", "minLength": 1 }, "active": { "type": "boolean" } },
        "additionalProperties": true
      },
...
    }
  }
}